+++ packages/mobile/package.json
@@ -31,6 +31,7 @@
   "devDependencies": {
     "@capacitor/cli": "5.5.1",
     "@metamask/detect-provider": "^2.0.0",
+    "@rainbow-me/rainbowkit": "^1.3.0",
     "@testing-library/dom": ">=7.21.4",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.0.0",
@@ -42,11 +43,14 @@
     "cypress": "^13.5.0",
     "eslint": "^8.35.0",
     "eslint-plugin-react": "^7.32.2",
+    "ethers": "^05.x",
     "jsdom": "^22.1.0",
     "terser": "^5.4.0",
     "typescript": "^5.1.6",
+    "viem": "^1.19.11",
     "vite": "5.0.4",
-    "vitest": "^0.34.6"
+    "vitest": "^0.34.6",
+    "wagmi": "^1.4.8"
   },
   "description": "An Ionic project"
 }
