+++ packages/mobile/src/App.tsx
@@ -31,12 +31,13 @@ const App: React.FC = () => (
     <IonMenu contentId="main-content">
         <IonHeader>
           <IonToolbar>
-            <IonTitle>Choose your role</IonTitle>
+            <IonTitle>
+                Profile
+            </IonTitle>
           </IonToolbar>
         </IonHeader>
         <IonContent className="ion-padding">
           <IonMenuToggle>
-            {/* <IonButton>Click to close the menu</IonButton> */}
             <IonButton expand="full" fill='clear' routerLink={`/home/proponent`}>Proponent</IonButton>
             <IonButton expand="full" fill='clear' routerLink={`/home/inversionist`}>Inversionist</IonButton>
             <IonButton expand="full" fill='clear' routerLink={`/home/validator`}>Validator</IonButton>
