+++ packages/mobile/vite.config.ts
@@ -4,9 +4,11 @@ import { defineConfig } from 'vite'
 
 // https://vitejs.dev/config/
 export default defineConfig({
+ build: {
+    target: 'es2020',
+  },
   plugins: [
     react(),
-    legacy()
   ],
   test: {
     globals: true,
